fix eslint and move statement card to a separate file
