+++ src/javascript/app_2/store/pages/statement/statement_model.js
